refactor: (#64) 컨테이너 크기 조절 및 그림 상태 관리 로직 훅 분리
